[ResponseAPI] Further polish message serialization and unit tests
